+++ vis/js/reducers/heading.js
@@ -18,7 +18,6 @@ const context = (state = {}, action) => {
         presetTitle: action.configObject.title,
         titleStyle: getTitleStyle(action.configObject),
         titleLabelType: getTitleLabelType(action.configObject),
-        // customTitle: action.configObject.custom_title,
         customTitle: action.contextObject.params.custom_title,
       };
     default:
