+++ vis/js/templates/modals/infomodal/subcomponents/StandardKMInfo.jsx
@@ -4,6 +4,8 @@ import { Modal } from "react-bootstrap";
 import AboutSoftware from "./AboutSoftware";
 import DataSource from "./DataSource";
 import { queryConcatenator } from "../../../../utils/data";
+import {getParameterValueFromLink} from "../../../../utils/getParamValueFromLink";
+
 
 const StandardKMInfo = ({
   serviceName,
@@ -88,29 +90,5 @@ const StandardKMInfo = ({
     // html template ends here
   );
 };
-
-function getParameterValueFromLink(parameterName) {
-  // Get the URL of the current page
-  const url = window.location.href;
-
-  // Parse the URL to extract the query parameters
-  const queryString = url.split('?')[1];
-  if (!queryString) {
-    return null; // No query parameters found
-  }
-
-  // Split the query string into individual parameters
-  const parameters = queryString.split('&');
-
-  // Loop through the parameters to find the one with the matching name
-  for (const parameter of parameters) {
-    const [name, value] = parameter.split('=');
-    if (name === parameterName) {
-      return decodeURIComponent(value); // Return the parameter value
-    }
-  }
-
-  return null; // Parameter not found
-}
 
 export default StandardKMInfo;
